refactor(template): isolate path discovery
Keep the generated entry point below complexity thresholds while preserving relocated-pipeline behavior.
